i64

+++ yt/yt/server/job_proxy/stderr_writer.cpp
@@ -154,7 +154,7 @@ NApi::TGetJobStderrResponse TStderrWriter::GetCurrentData(const NApi::TGetJobStd
     auto data = stringStream.Str();
     auto endOffset = TotalSize_;
     const i64 currentFirstAbsolutePos = TotalSize_ > static_cast<i64>(data.size()) ? TotalSize_ - data.size() : 0;
-    size_t dataBeginOffset;
+    i64 dataBeginOffset;
     if (offset >= 0) {
         if (offset >= currentFirstAbsolutePos) {
             dataBeginOffset = offset - currentFirstAbsolutePos;
